Simplify navigation components and consolidate link logic
