+++ apps/ffdweb-site/src/app/_components/Navigation/DesktopNavigation.tsx
@@ -1,8 +1,7 @@
 import { clsx } from 'clsx'
 
-import { NAV_LINKS, TOUCH_TARGET_NAV_LINK } from '../constants'
-
-import { DesktopLink } from './DesktopLink'
+import { NAV_LINKS, TOUCH_TARGET_NAV_LINK } from './constants'
+import { NavigationLink } from './NavigationLink'
 
 export function DesktopNavigation() {
   return (
@@ -14,7 +13,7 @@ export function DesktopNavigation() {
       )}
     >
       {NAV_LINKS.map((link) => (
-        <DesktopLink key={link.path} href={link.path} label={link.label} />
+        <NavigationLink key={link.path} href={link.path} label={link.label} />
       ))}
     </ul>
   )
