--- apps/ffdweb-site/src/app/_components/Navigation/MobileNavigation/MobileLink.tsx
@@ -1,34 +1,0 @@
-'use client'
-
-import Link, { type LinkProps } from 'next/link'
-import { usePathname } from 'next/navigation'
-
-import { clsx } from 'clsx'
-
-import { TOUCH_TARGET_NAV_LINK } from '../constants'
-
-type MobileLinkProps = {
-  label: string
-  href: LinkProps['href']
-  onClick?: LinkProps['onClick']
-}
-
-export function MobileLink({ href, label, ...rest }: MobileLinkProps) {
-  const pathname = usePathname()
-  const isActive = pathname.startsWith(href.toString())
-
-  return (
-    <Link
-      href={href}
-      aria-label={`Go to ${label} page`}
-      aria-current={isActive}
-      className={clsx(
-        TOUCH_TARGET_NAV_LINK.touchAreaPadding,
-        'focus:brand-outline aria-current:text-brand-primary-300 aria-current:bg-brand-primary-900 inline-block text-neutral-200 hover:bg-neutral-800',
-      )}
-      {...rest}
-    >
-      {label}
-    </Link>
-  )
-}
